# Huffman decode optimizations