--- a/src/ImageSharp/Formats/Jpeg/JpegConstants.cs
+++ b/src/ImageSharp/Formats/Jpeg/JpegConstants.cs
@@ -1,7 +1,8 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using System.Collections.Generic;
+using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder;
 
 namespace SixLabors.ImageSharp.Formats.Jpeg
 {
@@ -248,6 +249,11 @@ namespace SixLabors.ImageSharp.Formats.Jpeg
             /// The size of the huffman decoder register.
             /// </summary>
             public const int RegisterSize = 64;
+
+            /// <summary>
+            /// The minimum number of bits required by the <see cref="HuffmanScanBuffer"/>.
+            /// </summary>
+            public const int MinBits = 48;
 
             /// <summary>
             /// If the next Huffman code is no more than this number of bits, we can obtain its length