--- a/src/ImageSharp/Formats/Jpeg/JpegConstants.cs
+++ b/src/ImageSharp/Formats/Jpeg/JpegConstants.cs
@@ -256,6 +256,11 @@ namespace SixLabors.ImageSharp.Formats.Jpeg
             public const int FetchBits = 48;
 
             /// <summary>
+            /// The number of times to read the input stream when filling the <see cref="HuffmanScanBuffer"/> buffer.
+            /// </summary>
+            public const int FetchLoop = FetchBits / 8;
+
+            /// <summary>
             /// The minimum number of bits allowed before by the <see cref="HuffmanScanBuffer"/> before fetching.
             /// </summary>
             public const int MinBits = RegisterSize - FetchBits;