### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Some minor optimizations to the HuffmanScanDecoder plus a little cleanup. For baseline jpegs we're sitting at around 25-30% for the total processing time.

``` bash
Full decoder

|                         Method |            TestImage |       Mean |      Error |     StdDev | Ratio | RatioSD |     Gen 0 | Gen 1 | Gen 2 |   Allocated |
|------------------------------- |--------------------- |-----------:|-----------:|-----------:|------:|--------:|----------:|------:|------:|------------:|
| 'Decode Jpeg - System.Drawing' | Jpg/b(...)e.jpg [21] |   6.121 ms |   3.134 ms |  0.1718 ms |  1.00 |    0.00 |   46.8750 |     - |     - |   205.83 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/b(...)e.jpg [21] |  16.796 ms |   9.187 ms |  0.5036 ms |  2.74 |    0.03 |         - |     - |     - |    14.37 KB |
|                                |                      |            |            |            |       |         |           |       |       |             |
| 'Decode Jpeg - System.Drawing' | Jpg/b(...)f.jpg [28] |  16.034 ms |  19.142 ms |  1.0492 ms |  1.00 |    0.00 |  171.8750 |     - |     - |   757.04 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/b(...)f.jpg [28] |  38.926 ms |   4.144 ms |  0.2271 ms |  2.43 |    0.15 |         - |     - |     - |    15.35 KB |
|                                |                      |            |            |            |       |         |           |       |       |             |
| 'Decode Jpeg - System.Drawing' | Jpg/i(...)e.jpg [43] | 408.578 ms | 202.366 ms | 11.0924 ms |  1.00 |    0.00 | 1000.0000 |     - |     - |  7403.76 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/i(...)e.jpg [43] | 348.491 ms | 103.983 ms |  5.6997 ms |  0.85 |    0.04 |         - |     - |     - | 35177.02 KB |


The below benchmark is the result of returning an empty image before any iDCT or Color conversion. It appears those operations now take ~75%
of the decoding process.

|                         Method |            TestImage |       Mean |      Error |    StdDev | Ratio | RatioSD |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|------------------------------- |--------------------- |-----------:|-----------:|----------:|------:|--------:|----------:|------:|------:|-----------:|
| 'Decode Jpeg - System.Drawing' | Jpg/b(...)e.jpg [21] |   6.775 ms |   8.296 ms | 0.4547 ms |  1.00 |    0.00 |   46.8750 |     - |     - |  205.83 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/b(...)e.jpg [21] |   4.689 ms |   1.700 ms | 0.0932 ms |  0.69 |    0.06 |         - |     - |     - |   12.29 KB |
|                                |                      |            |            |           |       |         |           |       |       |            |
| 'Decode Jpeg - System.Drawing' | Jpg/b(...)f.jpg [28] |  18.201 ms |   3.556 ms | 0.1949 ms |  1.00 |    0.00 |  156.2500 |     - |     - |  757.04 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/b(...)f.jpg [28] |  14.439 ms |   4.345 ms | 0.2382 ms |  0.79 |    0.00 |         - |     - |     - |   12.29 KB |
|                                |                      |            |            |           |       |         |           |       |       |            |
| 'Decode Jpeg - System.Drawing' | Jpg/i(...)e.jpg [43] | 438.533 ms |  55.331 ms | 3.0329 ms |  1.00 |    0.00 | 1000.0000 |     - |     - | 7403.76 KB |
|     'Decode Jpeg - ImageSharp' | Jpg/i(...)e.jpg [43] | 229.262 ms | 100.169 ms | 5.4906 ms |  0.52 |    0.01 |         - |     - |     - | 35171.8 KB |
```

<!-- Thanks for contributing to ImageSharp! -->